Move IsStandard functions to policy
